+++ app/src/main/java/com/kprights/weatherapp/model/result/Sys.kt
@@ -9,6 +9,6 @@ data class Sys (
 		@SerializedName("type") var type : Int = 0,
 		@SerializedName("id") var id : Int = 0,
 		@SerializedName("country") var country : String = "",
-		@SerializedName("sunrise") var sunrise : Int = 0,
-		@SerializedName("sunset") var sunset : Int = 0
+		@SerializedName("sunrise") var sunrise : Long = 0L,
+		@SerializedName("sunset") var sunset : Long = 0L
 )
